Added easily adjustable Parameter & Label Heuristics

+++ src/de/tu_darmstadt/ke/seco/multilabelrulelearning/evaluation/boosting/LinLogBoosting.java
@@ -27,7 +27,7 @@ public class LinLogBoosting extends BoostingStrategy {
     }
 
     @Override
-    public double evaluate(MultiHeadRule rule, int numberOfLabelsInTheHead) {
+    public double evaluate(MultiHeadRule rule, double numberOfLabelsInTheHead) {
         this.numberOfLabelsInTheHead = numberOfLabelsInTheHead;
         rawRuleValue = rule.getRawRuleValue();
         boostedRuleValue = linLogValue();
