+++ src/de/tu_darmstadt/ke/seco/multilabelrulelearning/evaluation/boosting/LogAlphaBoosting.java
@@ -24,7 +24,7 @@ public class LogAlphaBoosting extends BoostingStrategy {
     }
 
     @Override
-    public double evaluate(MultiHeadRule rule, int numberOfLabelsInTheHead) {
+    public double evaluate(MultiHeadRule rule, double numberOfLabelsInTheHead) {
         this.numberOfLabelsInTheHead = numberOfLabelsInTheHead;
         rawRuleValue = rule.getRawRuleValue();
         boostedRuleValue = logValue();
